+++ info/source.c
@@ -234,7 +234,7 @@ static unsigned long ScanSourceFileNames
 }
 
 AspSourceLocation AspGetSourceLocation
-    (const AspSourceInfo *info, uint32_t pc)
+    (const AspSourceInfo *info, size_t pc)
 {
     /* Locate the applicable source info record. */
     const uint8_t *p = info->sourceInfos, *pp = p;
